checkpointing text package restructuring
